+++ src/Libraries/SmartStore.Core/Domain/Orders/Order.cs
@@ -427,6 +427,12 @@ namespace SmartStore.Core.Domain.Orders
 		[DataMember]
 		public DateTime UpdatedOnUtc { get; set; }
 
+		/// <summary>
+		/// Gets or sets the amount of remaing reward points
+		/// </summary>
+		[DataMember]
+		public int? RewardPointsRemaining { get; set; }
+
         #endregion
 
         #region Navigation properties