+++ src/Libraries/SmartStore.Data/Migrations/201407191010096_BetterReturnRequest.cs
@@ -9,6 +9,7 @@ namespace SmartStore.Data.Migrations
         public override void Up()
         {
 			AddColumn("dbo.Order", "UpdatedOnUtc", c => c.DateTime(nullable: false));
+			AddColumn("dbo.Order", "RewardPointsRemaining", c => c.Int());
 
 			try
 			{
@@ -48,12 +49,33 @@ namespace SmartStore.Data.Migrations
 				"Rücksendung erstellen");
 
 			builder.AddOrUpdate("Admin.ReturnRequests.Accept",
-				"Accept and update order",
-				"Akzeptieren und Auftrag anpassen");
+				"Accept",
+				"Akzeptieren");
 
-			builder.AddOrUpdate("Admin.ReturnRequests.Accept.Hint",
-				"Sets the status to accepted and updates quantities and totals of the order, same with the inventory and reward points.",
-				"Setzt den Status auf genehmigt und passt die Mengen und Summen des Auftrags, so wie Warenbestände und Bonuspunkte entsprechend an.");
+			builder.AddOrUpdate("Admin.ReturnRequests.Accept.Caption",
+				"Accept return request",
+				"Rücksendung akzeptieren");
+
+
+			builder.AddOrUpdate("Admin.Orders.OrderItem.Cancel.Info",
+				"Stock quantity old {0}, new {1}.<br />Reward points old {2}, new {3}.",
+				"Lagerbestand alt {0}, neu {1}.<br />Bonuspunkte alt {2}, neu {3}.");
+
+			builder.AddOrUpdate("Admin.Orders.OrderItem.Cancel.Caption",
+				"Remove product",
+				"Produkt entfernen");
+
+			builder.AddOrUpdate("Admin.Orders.OrderItem.Cancel.Fields.AdjustInventory",
+				"Adjust inventory",
+				"Lagerbestand anpassen",
+				"Determines whether to increase the stock quantity.",
+				"Legt fest, ob der Lagerbestand wieder erhöht werden soll.");
+
+			builder.AddOrUpdate("Admin.Orders.OrderItem.Cancel.Fields.ReduceRewardPoints",
+				"Adjust reward points",
+				"Bonuspunkte anpassen",
+				"Determines whether granted reward points should be deducted again proportionately.",
+				"Legt fest, ob gewährte Bonuspunkte wieder anteilig abgezogen werden sollen.");
 		}
     }
 }