+++ changelog.md
@@ -15,6 +15,7 @@
 * #393 Web API: Implement OData actions for simpler working with product attributes
 * #431 Web API: Add support for localized properties
 * ShippingByWeight: new settings to configure a small quantity surcharge
+* #216 Better return request support
 
 ###Improvements###
 * Task Scheduler: